Converts Chip 1 to Chip 2 in one fell swoop.
This commit provides a complete overhaul of chip 1.0.
It brings in fragments.js, removing binding, expressions, diff, and
observer from chip.
It converts chip to javascript from coffee script.
It uses the binding for the route tag which was handled manually
previously.
Tests and examples have been updated.
Going for [attr], (event), and {{expression}} binding. We’ll see if it
is too different for people, but it really enhances the readability of
the HTML as it is easy to find the bindings this way. Also trying
[class:some-classname]=“{{expr}}” for toggling a class on/off. Reads
well IMO.